backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==4.2.0
drf-yasg==1.21.7
psycopg2-binary==2.9.7
django-filter==23.3
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0